scripts/requirements.txt
supabase==2.3.4
httpx==0.24.1
httpcore==0.17.3
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
